move thumbnail back; use context
