+++ src/frontend/src/tables/part/PartThumbTable.tsx
@@ -20,12 +20,12 @@ import { useQuery } from '@tanstack/react-query';
 import type React from 'react';
 import { Suspense, useState } from 'react';
 
-import { Thumbnail } from '@lib/components/images/Thumbnail';
 import { ApiEndpoints } from '@lib/enums/ApiEndpoints';
 import { apiUrl } from '@lib/functions/Api';
 import { showNotification } from '@mantine/notifications';
 import { IconX } from '@tabler/icons-react';
 import { api } from '../../App';
+import { Thumbnail } from '../../components/images/Thumbnail';
 
 /**
  * Input props to table
